+++ src/commonMain/kotlin/internal/KeyPairManagerImpl.kt
@@ -23,6 +23,6 @@ internal expect class KeyPairManagerImpl() : KeyPairManager {
     override suspend fun generateNewKey(userId: Int, keyId: String): Failure.KeyManagement.GenerationFailed?
     override suspend fun retrievePublicKey(userId: Int, keyId: String): Xor<ByteArray, Failure.KeyManagement.KeyExtractionFailed>
     override suspend fun retrievePrivateKey(userId: Int, keyId: String): Xor<ByteArray, Failure.KeyManagement.KeyExtractionFailed>
-    override suspend fun findKeyIdFor(userId: Int): Xor<String, Failure.KeyManagement.KeyNotFound>
+    override suspend fun findKeyIdFor(userId: Long): Xor<String, Failure.KeyManagement.KeyNotFound>
     override suspend fun deleteKey(keyId: String): Xor<Unit, Failure.KeyManagement.KeyNotFound>
 }
